examples/requirements.txt
motorcortex-python==0.22.9
motorcortex-robot-control-python==3.1.2
pynng==0.7.4
